+++ src/periph_rfm69.h
@@ -48,7 +48,7 @@ RFMSend_t rfmSendBuffer(const uint8_t n, const uint8_t retries,
 /*! @brief Sets the RFM69's address
  *  @param [in] addr : address to set the RFM69
  */
-void rfmSetAddress(const uint16_t addr);
+void rfmSetAddress(const uint8_t addr);
 
 /*! @brief Set the AES key for encryption
  *  @param [in] aes : 16 character AES key. 0 disables encryption.
